requirements.txt
# Flask and extensions
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
Flask-Login>=0.6.3
Flask-Mail>=0.9.1
Flask-WTF>=1.2.1
Flask-Limiter>=3.5.0

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Authentication
Authlib>=1.3.0
Werkzeug>=3.0.0

# Solana
solders>=0.21.0
PyNaCl>=1.5.0
base58>=2.1.1

# HTTP requests
requests>=2.31.0

# Image processing and storage
cloudinary>=1.36.0
Pillow>=10.1.0

# HTML sanitization
bleach>=6.1.0

# Scheduler
APScheduler>=3.10.4

# Production server
gunicorn>=21.2.0

# Environment variables
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-flask>=1.3.0
